Implemented a CNN model using TensorFlow to classify car and bike images with high accuracy. Trained the model on a large dataset and optimized hyperparameters to achieve superior performance. Ready to deploy for real-world image recognition applications.
The commit message refers to a specific task that has been accomplished using TensorFlow, which is an open-source machine learning framework. In this task, a convolutional neural network (CNN) model has been developed to classify images of cars and bikes with high accuracy.

To achieve this, a large dataset of car and bike images was used to train the model. During the training process, the model's hyperparameters were optimized to ensure optimal performance. Once the model was trained, it was evaluated on a separate set of images to confirm its ability to accurately identify cars and bikes.

Finally, the developed model is ready to be deployed for real-world image recognition applications where it can be used to automate the process of identifying cars and bikes in images. The commit message signifies that the task has been successfully completed and is now ready for deployment.